+++ Sources/SwiftyMockyCLICore/Commands/MacOS/InspectionController.swift
