+++ Sources/SwiftyMockyCLICore/Utils/ProjectFile.swift
@@ -1,0 +1,15 @@
+import Foundation
+
+#if os(macOS)
+import xcodeproj
+
+extension PBXTarget: ProjectTarget { }
+#endif
+
+public protocol ProjectFile {
+    
+}
+
+public protocol ProjectTarget {
+    var name: String { get }
+}
