+++ Sources/SwiftyMockyCLICore/Utils/ShellOut+Extensions.swift
@@ -1,5 +1,4 @@
 import Foundation
-import Commander
 
 class ProxyFileHandle: FileHandle {
     override func write(_ data: Data) {
